inform schema path
