+++ packages/global/support/user/constant.ts
@@ -1,0 +1,9 @@
+export enum InformTypeEnum {
+  system = 'system'
+}
+
+export const InformTypeMap = {
+  [InformTypeEnum.system]: {
+    label: '系统通知'
+  }
+};
