+++ projects/app/src/pages/api/user/inform/send.ts
@@ -1,11 +1,15 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { jsonRes } from '@/service/response';
-import { Inform, connectToDatabase } from '@/service/mongo';
+import { connectToDatabase } from '@/service/mongo';
 import { authUser } from '@fastgpt/service/support/user/auth';
-import { InformTypeEnum } from '@/constants/user';
 import { startSendInform } from '@/service/events/sendInform';
-import { MongoUser } from '@fastgpt/service/support/user/schema';
+import { MongoUserInform } from '@fastgpt/service/support/user/inform/schema';
+import { InformTypeEnum } from '@fastgpt/global/support/user/constant';
+import {
+  sendInform2AllUser,
+  sendInform2OneUser
+} from '@fastgpt/service/support/user/inform/controller';
 
 export type Props = {
   type: `${InformTypeEnum}`;
@@ -38,39 +42,13 @@ export async function sendInform({ type, title, content, userId }: Props) {
 
   try {
     if (userId) {
-      global.sendInformQueue.push(async () => {
-        // skip it if have same inform within 5 minutes
-        const inform = await Inform.findOne({
-          type,
-          title,
-          content,
-          userId,
-          time: { $gte: new Date(Date.now() - 5 * 60 * 1000) }
-        });
-
-        if (inform) return;
-
-        await Inform.create({
-          type,
-          title,
-          content,
-          userId
-        });
-      });
+      global.sendInformQueue.push(async () => sendInform2OneUser({ type, title, content, userId }));
       startSendInform();
       return;
     }
 
     // send to all user
-    const users = await MongoUser.find({}, '_id');
-    await Inform.insertMany(
-      users.map(({ _id }) => ({
-        type,
-        title,
-        content,
-        userId: _id
-      }))
-    );
+    sendInform2AllUser({ type, title, content });
   } catch (error) {
     console.log('send inform error', error);
   }
