+++ projects/app/src/service/mongo.ts
@@ -60,4 +60,3 @@ export * from './models/chat';
 export * from './models/chatItem';
 export * from './models/app';
 export * from './common/bill/schema';
-export * from './models/inform';
